Display GITHUB_SHA on the index page if it exists
We can optionally pass a GITHUB_SHA_ARG build argument to the
Docker container while we're building it.

If we don't, no problem.
If we do, we will see the commit referenced on the index page.

This is useful to us because we want to make sure that automatically
deploying our code, we are actually getting new versions of it, which
is hard to tell if there haven't been any layout or content changes.